Remove agency client from libvcx-core tests
