+++ aries_vcx/src/utils/devsetup.rs
@@ -6,8 +6,6 @@ use std::sync::{Arc, Once};
 
 use chrono::{DateTime, Duration, Utc};
 
-use agency_client::agency_client::AgencyClient;
-use agency_client::configuration::AgentProvisionConfig;
 use agency_client::testing::mocking::{enable_agency_mocks, AgencyMockDecrypted};
 use aries_vcx_core::global::settings::{
     disable_indy_mocks as disable_indy_mocks_core, enable_indy_mocks as enable_indy_mocks_core,
@@ -15,9 +13,8 @@ use aries_vcx_core::global::settings::{
 };
 use aries_vcx_core::ledger::base_ledger::{AnoncredsLedgerRead, AnoncredsLedgerWrite, IndyLedgerRead, IndyLedgerWrite};
 use aries_vcx_core::ledger::indy::pool::test_utils::{create_testpool_genesis_txn_file, get_temp_file_path};
-use aries_vcx_core::wallet::base_wallet::BaseWallet;
 use aries_vcx_core::wallet::indy::did_mocks::DidMocks;
-use aries_vcx_core::wallet::indy::wallet::{create_and_open_wallet, create_and_store_my_did, wallet_configure_issuer};
+use aries_vcx_core::wallet::indy::wallet::{create_and_open_wallet, create_and_store_my_did};
 use aries_vcx_core::wallet::indy::{IndySdkWallet, WalletConfig};
 use aries_vcx_core::WalletHandle;
 
@@ -34,7 +31,6 @@ use crate::global::settings::{
 use crate::global::settings::{init_issuer_config, reset_config_values_ariesvcx};
 use crate::utils::constants::{POOL1_TXN, TRUSTEE_SEED};
 use crate::utils::file::write_file;
-use crate::utils::provision::provision_cloud_agent;
 use crate::utils::test_logger::LibvcxDefaultLogger;
 
 #[macro_export]
@@ -134,39 +130,6 @@ impl Drop for SetupMocks {
     fn drop(&mut self) {
         reset_global_state();
     }
-}
-
-// todo: we move to libvcx?
-pub async fn dev_setup_issuer_wallet_and_agency_client() -> (String, WalletHandle, AgencyClient) {
-    let enterprise_seed = "000000000000000000000000Trustee1";
-    let config_wallet = WalletConfig {
-        wallet_name: format!("wallet_{}", uuid::Uuid::new_v4()),
-        wallet_key: settings::DEFAULT_WALLET_KEY.into(),
-        wallet_key_derivation: settings::WALLET_KDF_RAW.into(),
-        wallet_type: None,
-        storage_config: None,
-        storage_credentials: None,
-        rekey: None,
-        rekey_derivation_method: None,
-    };
-    let config_provision_agent = AgentProvisionConfig {
-        agency_did: AGENCY_DID.to_string(),
-        agency_verkey: AGENCY_VERKEY.to_string(),
-        agency_endpoint: AGENCY_ENDPOINT.parse().expect("valid url"),
-        agent_seed: None,
-    };
-    let wallet_handle = create_and_open_wallet(&config_wallet).await.unwrap();
-    let config_issuer = wallet_configure_issuer(wallet_handle, enterprise_seed).await.unwrap();
-    init_issuer_config(&config_issuer.institution_did).unwrap();
-    let mut agency_client = AgencyClient::new();
-
-    let wallet: Arc<dyn BaseWallet> = Arc::new(IndySdkWallet::new(wallet_handle));
-
-    provision_cloud_agent(&mut agency_client, wallet, &config_provision_agent)
-        .await
-        .unwrap();
-
-    (config_issuer.institution_did, wallet_handle, agency_client)
 }
 
 pub async fn dev_setup_wallet_indy(key_seed: &str) -> (String, WalletHandle) {
